backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-generativeai==0.5.4
firebase-admin==6.5.0
httpx==0.27.0
python-multipart==0.0.9
Pillow==10.3.0
python-dotenv==1.0.1
pydantic==2.7.1
pytest==8.2.0
pytest-asyncio==0.23.7
